+++ tests/integration/command_mode_test.py
@@ -1,5 +1,6 @@
 import os
 import time
+from typing import Final
 
 from setup import setup
 from setup import TmuxRunner
@@ -415,3 +416,14 @@ def test_run_shell_cmd_stderr(r: TmuxRunner):
     msg: str = r.color_screenshot()[-1]
     assert "hello" in msg
     assert "255;0;0" in msg  # red text
+
+
+@setup("tests/fixture/lorem_ipsum.txt")
+def test_list_marks(r: TmuxRunner):
+    r.type_str("ma")
+    r.iris_cmd("lm")
+
+    title_line: Final[str] = r.lines()[-11]
+    assert "Marks" in title_line
+    assert "lorem_ipsum.txt" in title_line
+    assert "a | 0:0 " in r.lines()[-10]