+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Disabled continuous deployment in GHA. See `.github/workflows/build_artifact.yml`
 for details
 * `M<char>` now marks the current row to the given letter on the keyboard
+* `;lm` now lists all marks in the current file in an overlay
 * `--version` now states `Unknown` if the git tag/hash cannot be found
 
 * Resolve crash when user tries to switch top or bottom line outside of scope